Refactor rdupes for best practices and performance
- Modularized the codebase into `types`, `hashing`, `traversal`, and `finder`.
- Introduced `PathGroup` and `FileKey` to correctly handle hardlinks.
- Optimized hashing for small files (size <= 4096).
- Replaced `async-fs` and `futures-lite` with `tokio::fs` and `tokio::io`.
- Replaced `tokio-task-tracker` with `tokio::task::JoinSet`.
- Improved error handling by replacing `unwrap()` and `expect()` with `Result`.
- Added `serde` as a dependency for `PathGroup` serialization.
