+++ Cargo.toml
@@ -8,22 +8,20 @@ description = "A rust port of pydupes. Super fast."
 
 
 [dependencies]
-async-fs = "2.1"
 blake3 = { version = "1.5", features = ["mmap"] }
 ciborium = "0.2.2"
 clap = { version = "4.5", features = ["derive"] }
 clio = { version = "0.3.5", features = ["clap-parse"] }
 closure = "0.3"
 futures = "0.3"
-futures-lite = "2.3"
 humansize = "2.1"
 kdam = "0.5"
 log = { features = ['release_max_level_info'], version = "0.4" }
 mimalloc = { version = "0.1", default-features = false }
+serde = { version = "1.0.228", features = ["derive"] }
 simple_logger = { features = ['stderr'], version = "4.3" }
 tinyvec = { version = "1.8.0", features = ["alloc", "serde"] }
 tokio = { version = "1", features = ["full"] }
-tokio-task-tracker = "1.3"
 tokio-util = "0.7"
 walkdir = "2.5"
